test: Move clicking helper functions to `wizardTestUtils`
As clicking helper functions are used only in the context of Wizard, this moves them to an appropriate place.
