+++ src/test/Components/CreateImageWizard/steps/Details/Details.test.tsx
@@ -4,7 +4,7 @@ import { userEvent } from '@testing-library/user-event';
 import { CREATE_BLUEPRINT, EDIT_BLUEPRINT } from '../../../../../constants';
 import { mockBlueprintIds } from '../../../../fixtures/blueprints';
 import { detailsCreateBlueprintRequest } from '../../../../fixtures/editMode';
-import { clickNext, getNextButton } from '../../../../testUtils';
+import { clickNext, getNextButton } from '../../wizardTestUtils';
 import {
   blueprintRequest,
   clickRegisterLater,
