+++ src/test/Components/CreateImageWizard/steps/Repositories/Repositories.test.tsx
@@ -13,7 +13,7 @@ import {
   expectedPayloadRepositories,
   repositoriesCreateBlueprintRequest,
 } from '../../../../fixtures/editMode';
-import { clickNext } from '../../../../testUtils';
+import { clickNext } from '../../wizardTestUtils';
 import {
   blueprintRequest,
   clickRegisterLater,
