+++ src/test/Components/CreateImageWizard/steps/TargetEnvironment/GCPTarget.test.tsx
@@ -10,7 +10,7 @@ import {
 } from '../../../../../store/imageBuilderApi';
 import { mockBlueprintIds } from '../../../../fixtures/blueprints';
 import { gcpCreateBlueprintRequest } from '../../../../fixtures/editMode';
-import { clickBack, clickNext } from '../../../../testUtils';
+import { clickBack, clickNext } from '../../wizardTestUtils';
 import {
   blueprintRequest,
   clickRegisterLater,
